+++ deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/conf/ocnn/OCNNOutputLayer.java
@@ -99,9 +99,9 @@ public class OCNNOutputLayer extends BaseOutputLayer {
     }
 
     @Override
-    public int getNOut() {
+    public long getNOut() {
         //we don't change number of outputs here
-        return 1;
+        return 1L;
     }
 
     @Override
